fix: rename hook folder to hooks
